refactor(dashboard): extend lenscommon helpers for Lens converters
Centralize number-format normalization, ES|QL format unmarshaling, XY plan JSON
comparison helpers, partition group-by JSON extraction, optional TF snapshot
mappers, and ES|QL partition nested schemas. Wire dashboard callers (XY state
alignment, waffle/mosaic/treemap formatting, partition helpers) and trim unused
wrappers.
