+++ internal/kibana/dashboard/models.go
@@ -36,9 +36,6 @@ import (
 // jsonNullString is the JSON encoding of null; json.Marshal uses it for unset union/API fields.
 const jsonNullString = "null"
 
-// defaultNumberFormatJSON is the fallback Kibana number format used when the API omits format.
-const defaultNumberFormatJSON = `{"type":"number"}`
-
 // populateFromAPI populates the Terraform model from the API response
 func dashboardPopulateFromAPI(ctx context.Context, m *models.DashboardModel, resp *kbapi.GetDashboardsIdResponse, dashboardID string, spaceID string) diag.Diagnostics {
 	var diags diag.Diagnostics
