+++ internal/kibana/dashboard/models_lens_chart_helpers.go
@@ -19,7 +19,6 @@ package dashboard
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/elastic/terraform-provider-elasticstack/generated/kbapi"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/lenscommon"
@@ -68,31 +67,14 @@ func preservePriorJSONWithDefaultsIfEquivalent[T any](ctx context.Context, prior
 // JSON is the ES|QL ("esql") or table ("table") dataset shape. Both map to the
 // ES|QL API variant of a chart panel. Returns false on marshal/unmarshal error.
 func lensDataSourceIsESQLOrTable(body []byte, err error) bool {
-	if err != nil {
-		return false
-	}
-	var ds struct {
-		Type string `json:"type"`
-	}
-	if err := json.Unmarshal(body, &ds); err != nil {
-		return false
-	}
-	return ds.Type == legacyMetricDatasetTypeESQL || ds.Type == legacyMetricDatasetTypeTable
+	return lenscommon.LensDataSourceIsESQLOrTable(body, err)
 }
 
 // lensESQLNumberFormatJSONFromAPI marshals a Lens ES|QL dimension `format` union
 // value to a normalized Terraform string. Empty or null JSON is replaced with the
 // default number-format payload so Terraform state matches what Kibana echoes.
 func lensESQLNumberFormatJSONFromAPI(format any, errLabel string, diags *diag.Diagnostics) (jsontypes.Normalized, bool) {
-	bytes, err := json.Marshal(format)
-	if err != nil {
-		diags.AddError("Failed to marshal "+errLabel, err.Error())
-		return jsontypes.Normalized{}, false
-	}
-	if len(bytes) == 0 || string(bytes) == jsonNullString {
-		bytes = []byte(defaultNumberFormatJSON)
-	}
-	return jsontypes.NewNormalizedValue(normalizeKibanaLensNumberFormatJSONString(string(bytes))), true
+	return lenscommon.LensESQLNumberFormatJSONFromAPI(format, errLabel, diags)
 }
 
 // lensQueryESQLMode returns whether a Lens chart's optional `query` object selects
