+++ internal/kibana/dashboard/panel/lensregionmap/converter.go
@@ -63,7 +63,8 @@ func (converter) SchemaAttribute() schema.Attribute {
 		Required:            true,
 	}
 	attrs["region_json"] = schema.StringAttribute{
-		MarkdownDescription: "Region configuration as JSON. For ES|QL, this defines the region column and EMS join. For standard mode, this defines the bucket operation (terms, histogram, range, filters) and optional EMS settings.",
+		MarkdownDescription: "Region configuration as JSON. For ES|QL, this defines the region column and EMS join. " +
+			"For standard mode, this defines the bucket operation (terms, histogram, range, filters) and optional EMS settings.",
 		CustomType:          jsontypes.NormalizedType{},
 		Required:            true,
 	}
